requirements.txt
Fabric3==1.13.1.post1
pyquery==1.2.17
requests==2.13.0